refactor(sandbox): move ls under a files namespace, restore sandbox ls alias
`bunny sandbox files list` (alias: ls) now houses the file listing,
mirroring `bunny storage files`, and gives future file ops (rm, mv) a
home. `bunny sandbox ls` is once again an alias of `sandbox list`.
