+++ .changeset/sandbox-ls-and-exec-timeout.md
@@ -2,4 +2,4 @@
 "@bunny.net/cli": minor
 ---
 
-feat(sandbox): add `bunny sandbox ls` to list files in a sandbox directory over SFTP (bare name lists `/workplace`, or `<sandbox>:<path>`), and `--timeout` on `bunny sandbox exec` to close the SSH connection and exit 124 after N seconds. The `ls` alias of `bunny sandbox list` now refers to the file listing; use `bunny sandbox list` to list sandboxes.
+feat(sandbox): add `bunny sandbox files list` (alias: `ls`) to list files in a sandbox directory over SFTP (bare name lists `/workplace`, or `<sandbox>:<path>`), and `--timeout` on `bunny sandbox exec` to close the SSH connection and exit 124 after N seconds.
